% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASSIGN_KIR3DL1}
\alias{ASSIGN_KIR3DL1}
\title{KIR3DL1/HLA-B Subtypes}
\format{\describe{
  \item{Allele}{KIR3DL1 allele}
  \item{Class}{Allele classification. 0 means unclassified}
}}
\source{
Boudreau et al. Journal of Clinical Oncology 2017 35:20, 2268-2278
}
\usage{
ASSIGN_KIR3DL1
}
\description{
KIR3DL1/HLA-B Subtypes Govern Acute Myelogenous Leukemia Relapse After Hematopoietic Cell Transplantation, Boudreau et al.
}
\keyword{datasets}
