% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafolder_check.R
\name{datafolder_update}
\alias{datafolder_update}
\title{Save data folder content list to sync to git}
\usage{
datafolder_update()
}
\description{
These functions have been moved to a separate package, see \link{https://github.com/bosefalk/datafolder}
}
