% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLA_B_class_data}
\alias{HLA_B_class_data}
\title{HLA-B Bw4/Bw6-80I/T ligand classification database}
\format{
\describe{
  \item{Allele}{HLA-B allele, in format "B*01:02:01:04"}
  \item{Predicted Ligand}{Predicted Ligand, "Bw6", "Bw4 - 80I" or "Bw4 - 80T"}
}
}
\source{
\url{https://www.ebi.ac.uk/cgi-bin/ipd/kir/retrieve_ligands.cgi?B}
}
\usage{
HLA_B_class_data
}
\description{
Classification database of HLA-B Ligands from Alleles, from Immuno Polymorphism Database
}
\seealso{
\code{\link{HLA_B_class_load}}, \code{\link{HLA_B_classification}}
}
\keyword{datasets}
