% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_B_class_load}
\alias{HLA_B_class_load}
\title{Load HLA-B Ligand Classification}
\usage{
HLA_B_class_load(fields = 2)
}
\arguments{
\item{fields}{Number of fields in the allele to keep (see \code{\link{shorten_allele}}).
Default is to keep the first two fields (i.e. 02:05)}
}
\value{
A data.frame with columns "Allele" and "Class"
}
\description{
Loads the HLA-B ligand classification database (Bw6, Bw4 - 80T, Bw4 - 80I classes), and keeps the specified first number
of allele fields / characters. Removes all "Unclassified" entries in the source data.
}
\examples{
# Default with two fields
head(HLA_B_class_load())
# With three fields
head(HLA_B_class_load(fields = 3))
}
\seealso{
\code{\link{HLA_B_class_data}}
}
