% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_B_classification}
\alias{HLA_B_classification}
\title{Classify overall group of HLA-B alleles}
\usage{
HLA_B_classification(allele_b1, allele_b2, fields = 2,
  HLA_B_class = NULL)
}
\arguments{
\item{allele_b1}{Allele, as string}

\item{allele_b2}{Allele, as string}

\item{fields}{Number of fields from allele string reference document to use, should be same as number of fields in \code{allel_b1}.
For example, if allel_b1 = "44:180", fields should be 2 (the default value). This does not actually apply
any string manipulation to \code{allele_b1} & \code{b2}, which needs to be done before passing to this function.}

\item{HLA_B_class}{optional, dataframe from \code{\link{HLA_B_class_load}}, if not supplied this will be called within the function
This option is here to pre-load this reference data, in case where you need to apply this function across a large number of subjects.}
}
\value{
One of "Bw6", "Bw4 - 80T", "Bw4 - 80I", or \code{NA_character_}, as a string
}
\description{
Takes two HLA-B alleles, classifies them individually into Bw6, Bw4 - 80I or Bw4 - 80T using \code{\link{HLA_Classification}} and
\code{\link{HLA_B_class_data}}, and returns the overall group. If either allele is 80I that's the overall group, if not then if either is 80T that's the group,
otherwise Bw6.If either allel classification is unknown, returns NA.
}
\examples{
HLA_B_classification("44:180", "07:02")
# For when you need to optimize execution time, pre-load the reference database
referencedata <- HLA_B_class_load()
HLA_B_classification("44:180", "07:02", HLA_B_class = referencedata)
}
\seealso{
\code{\link{HLA_Classification}}, \code{\link{HLA_B_class_data}}
}
