% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HLA_C_class_data}
\alias{HLA_C_class_data}
\title{HLA-C C1/C2 ligand classification database}
\format{\describe{
  \item{Allele}{HLA-C allele, in format "C*01:02:01:04"}
  \item{Predicted Ligand}{Predicted Ligand, "C1" or "C2"}
}}
\source{
\url{https://www.ebi.ac.uk/cgi-bin/ipd/kir/retrieve_ligands.cgi?C}
}
\usage{
HLA_C_class_data
}
\description{
Classification database of HLA-C Ligands from Alleles, from Immuno Polymorphism Database
}
\seealso{
\code{\link{HLA_C_class_load}}, \code{\link{HLA_C_classification}}
}
\keyword{datasets}
