% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_C_class_load}
\alias{HLA_C_class_load}
\title{Load HLA-C Ligand Classification}
\usage{
HLA_C_class_load(nchar = 5)
}
\arguments{
\item{nchar}{Number of characters in the allele to keep (after removing the inital "C*", but including ":").
Default is \code{nchar} = 5, which gives the first two fields (i.e. 02:05)}
}
\value{
A data.frame with columns "Allele" and "Predicted Ligand"
}
\description{
Loads the HLA-C C1/C2 ligand classification database, and keeps the specified first number
of allele fields / characters. Removes all "Unclassified" entries in the source data.
}
\examples{
# Default with two fields
head(HLA_C_class_load())
# Only the first field
head(HLA_C_class_load(nchar = 2))
}
\seealso{
\code{\link{HLA_C_class_data}}
}
