% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_C_classification}
\alias{HLA_C_classification}
\title{Classify HLA-C alleles into C1 & C2 groups}
\usage{
HLA_C_classification(allele_c1, allele_c2, nchar = 5)
}
\arguments{
\item{nchar}{Number of characters from allele string reference document to use, should be same as length as \code{allel_c1}.
For example, if allel_c1 = "02:03", nchar should be 5. Default value is 5, for the first two fields.}

\item{allel_c1}{Allele, as string}

\item{allel_c2}{Allele, as string - same length as \code{allel_c1}}
}
\value{
One of three C1/C2 groups, as string
}
\description{
Takes two HLA-C alleles, classifies them individually into C1 or C2 using \code{\link{HLA_Classification}} and \code{\link{HLA_C_class_data}},
and returns the joint C1/C1, C1/C2 or C2/C2 class. If either allel classification is unknown, returns NA.
}
\examples{
HLA_C_classification("01:02", "01:AWFCH")
}
\seealso{
\code{\link{HLA_Classification}}, \code{\link{HLA_C_class_data}}, \code{\link{NMDP}}
}
