% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_Classification}
\alias{HLA_Classification}
\title{Classify one HLA allele}
\usage{
HLA_Classification(allele, HLA_x_class)
}
\arguments{
\item{allele}{An HLA allele string vector, including ":" characers but not the initial "C*"}

\item{HLA_x_class}{HLA classification reference data.frame (such as constructed by for example \code{HLA_C_class_load()}), must
contain columns Allele (where the number of fields is the same as in \code{allele}) and Class.}
}
\value{
String vector with classification group from reference document, NA if unknown
}
\description{
Classifies an allele according to a reference dataframe. If an NMDP code is present, class is given if all possible
alleles are the same class, otherwise returns NA
}
\examples{

dat <- data.frame(HLA_C = c("01:02", "01:AWFCH"), HLA_B = c("07:02", NA),stringsAsFactors = FALSE)
HLA_C_class <- HLA_C_class_load()
HLA_Classification(dat$HLA_C, HLA_C_class)
HLA_B_class <- HLA_B_class_load()
HLA_Classification(dat$HLA_B, HLA_B_class)
}
\seealso{
\code{\link{HLA_C_classification}}, \code{\link{HLA_B_classification}},
\code{\link{HLA_C_class_load}}, \code{\link{HLA_B_class_load}}, \code{\link{NMDP}}
}
