% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR3DL1_HLA_B_inhibiting}
\alias{KIR3DL1_HLA_B_inhibiting}
\title{Classifies KIR3DL1 and HLA-B into strong/weak/no inhibiting}
\usage{
KIR3DL1_HLA_B_inhibiting(KIRD3L1_assignment, HLA_B_overall,
  three_levels = FALSE)
}
\arguments{
\item{KIRD3L1_assignment}{String, KIR3DL1 assignment in fromat "KIR3DL1-H", from \code{\link{KIR3DL1_3DS1_assignment}}}

\item{HLA_B_overall}{String with HLA-B Bw classification in format "Bw4 - 80T", from \code{\link{HLA_B_classification}}}

\item{three_levels}{If set to TRUE, returns separate results for "weak" and "noninhibiting". Default FALSE}
}
\value{
String with "Strong inhibiting", "Weak inhibiting/noninhibiting" or "unknown", unless three_levels
is set to TRUE in which case "Weak inhibiting" and "noninhibiting" are split.
}
\description{
Classifies KIR3DL1 and HLA-B into strong/weak/no inhibiting
}
\examples{
KIR3DL1_HLA_B_inhibiting("KIR3DL1-L", "Bw4 - 80T")
KIR3DL1_HLA_B_inhibiting("KIR3DL1-L", "Bw4 - 80I")
KIR3DL1_HLA_B_inhibiting("KIR3DL1-L", "Bw4 - 80I", three_levels = TRUE)
KIR3DL1_HLA_B_inhibiting("unknown", "Bw4 - 80I")
}
