% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR_det_GCN}
\alias{KIR_det_GCN}
\title{Determine KIR Gene Copy Number}
\usage{
KIR_det_GCN(kir_string, return_numeric = FALSE)
}
\arguments{
\item{kir_string}{String with KIR info with the first field only (three characters), example "004+003|006+010"}

\item{return_numeric}{If set to TRUE, returns numeric GCN instead of character, sets "x" to NA and 2|3 to 2.5 etc}
}
\value{
Default a string with "x" if "POS", "0" if NEG, NA if NA or empty string, and otherwise the GCN. If \code{return_numeric = TRUE}
instead returns a numeric value, where x results are NA and "2|3" are 2.5
}
\description{
Determine KIR Gene Copy Number
}
\examples{

str(KIR_det_GCN("004+003|006+010"))
str(KIR_det_GCN("004+003|006+010", return_numeric = TRUE))
KIR_det_GCN("POS")
KIR_det_GCN("POS", return_numeric = TRUE)
}
