% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{KIR_first_field}
\alias{KIR_first_field}
\title{Reduce KIR string to first field}
\usage{
KIR_first_field(kir_string)
}
\arguments{
\item{kir_string}{KIR string vector with /, + and | characters}
}
\value{
String vector with first KIR field only, removing duplicates
}
\description{
Reduce a long KIR string (i.e. "0010101/0020102+0020103/0020104|0030105/0030106/0030107") to keep only
the first field, resulting in "001/002+002|003". Keeps all non-numeric entries intact (i.e. "NEG" or "POS")
}
\examples{
dat <- data.frame(kirstring = c("0010101/0020102+0020103/0020104|0030105/0030106/0030107", "001/002", "NEG", "0010203+0020304|NEG", "POS", "", NA), stringsAsFactors = FALSE)
KIR_first_field(dat$kirstring)
}
