% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR_haplotype_Tel}
\alias{KIR_haplotype_Tel}
\title{KIR Telomeric Haplotype}
\usage{
KIR_haplotype_Tel(
  df,
  KIR_3DL1col,
  KIR_2DS4col,
  KIR_3DS1col,
  KIR_2DS1col,
  present_cols = FALSE
)
}
\arguments{
\item{df}{data.frame with KIR alleles string as columns}

\item{KIR_3DL1col}{string, name of 3DL1 column}

\item{KIR_2DS4col}{string, name of 2DS4 column}

\item{KIR_3DS1col}{string, name of 3DS1 column}

\item{KIR_2DS1col}{string, name of 2DS1 column}

\item{present_cols}{Default is FALSE which assumes the KIR columns contains the allele string ("001", "NEG" etc), and the presence / absence is calculated inside the function using \code{\link{KIR_present}}. If set to TRUE, the KIR columns already contain TRUE/FALSE for presence/absence for all KIRs.}
}
\value{
string vector with haplotype ("A/A", "A/B", "B/B"), NA if any input KIR is missing, and "unknown" otherwise.
}
\description{
Assigns the Telomeric haplotype (A/A, A/B or B/B) given a set of KIRs
}
\examples{
dat <- data.frame(kir_3DL1 = c("001", "NEG", NA, "001|002+003"), 
   kir_2DS4 = c("001", "001", "001", "001"), 
   kir_3DS1 = c("NEG", "001", "NEG", "NEG"),
   kir_2DS1 = c("001", "001", "NEG", "NEG"),
   stringsAsFactors = FALSE)
KIR_haplotype_Tel(dat, "kir_3DL1", "kir_2DS4", "kir_3DS1", "kir_2DS1")
   
dat_pres <- data.frame(kir_3DL1 = c(TRUE, FALSE, NA, TRUE), 
    kir_2DS4 = c(TRUE, TRUE, TRUE, TRUE), 
    kir_3DS1 = c(FALSE, TRUE, FALSE, FALSE),
    kir_2DS1 = c(TRUE, TRUE, FALSE, FALSE),
    stringsAsFactors = FALSE)
KIR_haplotype_Tel(dat_pres, "kir_3DL1", "kir_2DS4", "kir_3DS1", "kir_2DS1", present_cols = TRUE)
}
\seealso{
\code{\link{KIR_haplotype_Cen}}
}
