% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR_present}
\alias{KIR_present}
\title{KIR present/absent}
\usage{
KIR_present(string)
}
\arguments{
\item{string}{KIR string vector}
}
\value{
logical vector, TRUE if the KIR is present, FALSE if it's not present, NA if passed an empty/NA string
}
\description{
Determine whether the KIR is present, given a KIR string. Assumes a string with only "NEG" is absent, any other string not including
"NEG" is present, if missing or some combination together with "NEG" return NA - see examples.
}
\examples{
dat <- data.frame(kirstring = c("003/034+003/034", "NEG", "", NA, "003|NEG"), stringsAsFactors = FALSE)
KIR_present(dat$kirstring)
}
