% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Supertype_HLA_lookup}
\alias{Supertype_HLA_lookup}
\title{Superclass HLA lookup}
\format{
\describe{
  \item{Allele}{HLA allele}
  \item{Supertype}{Supertype}
  \item{HLA}{A for HLA-A, B for HLA-B allele}
}
}
\source{
\url{http://www.biomedcentral.com/1471-2172/9/1}
}
\usage{
Supertype_HLA_lookup
}
\description{
Superclass HLA allele assignments, column HLA equal to "A" for HLA-A, "B" for HLA-B
}
\keyword{datasets}
