% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafolder_check.R
\name{datafolder_check}
\alias{datafolder_check}
\title{Check local data folder is up-to-date}
\usage{
datafolder_check(stop_on_error = TRUE)
}
\description{
These functions have been moved to a separate package, see \link{https://github.com/bosefalk/datafolder}
}
