% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_C_classification}
\alias{HLA_C_classification}
\title{Classify HLA-C alleles into C1 & C2 groups}
\usage{
HLA_C_classification(allele_c1, allele_c2, fields = 2,
  HLA_C_class = NULL)
}
\arguments{
\item{fields}{Number of fields from allele string reference document to use, should be same as number of fields in \code{allel_c1}.
For example, if allel_c1 = "02:03", fields should be 2 (the default value). This does not actually apply
any string manipulation to \code{allele_c1} & \code{c2}, which needs to be done before passing to this function.}

\item{HLA_C_class}{optional, dataframe from \code{\link{HLA_C_class_load}}, if not supplied this will be called within the function
This option is here to pre-load this reference data, in case where you need to apply this function across a large number of subjects.}

\item{allel_c1}{Allele, as string}

\item{allel_c2}{Allele, as string}
}
\value{
One of "C1/C1", "C1/C2", "C2/C2", or \code{NA_character_}, as a string
}
\description{
Takes two HLA-C alleles, classifies them individually into C1 or C2 using \code{\link{HLA_Classification}} and \code{\link{HLA_C_class_data}},
and returns the joint C1/C1, C1/C2 or C2/C2 class. If either allel classification is unknown, returns NA.
}
\details{
There are only three outcome classes, C1/C1, C1/C2 and C2/C2. The location of each C-class before or after the "/" does not
match the input \code{allele_c1} and \code{_c2} - in other words if \code{allele_c1 = "C2"} and \code{allele_c2 = "C1"} the result is still C1/C2
}
\examples{
HLA_C_classification("01:02", "01:AWFCH")

# For when you need to optimize execution time, pre-load the reference database
referencedata <- HLA_C_class_load()
HLA_C_classification("01:02", "01:AWFCH", HLA_C_class = referencedata)
}
\seealso{
\code{\link{HLA_Classification}}, \code{\link{HLA_C_class_data}}
}
