% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_C_classification}
\alias{HLA_C_classification}
\title{Classify HLA-C alleles into C1 & C2 groups}
\usage{
HLA_C_classification(allele_c1, allele_c2, fields = 2, HLA_C_class = NULL)
}
\arguments{
\item{fields}{Number of fields from allele string reference document to use, should be same as number of fields in \code{allel_c1}.
For example, if allel_c1 = "02:03", fields should be 2 (the default value). This does not actually apply
any string manipulation to \code{allele_c1} & \code{allele_c2}, which needs to be done before passing to this function, 
using \code{\link{shorten_allel}}}

\item{HLA_C_class}{deprecated, used to pre-load classification data before this function was vectorized}

\item{allel_c1}{Allele column, as string}

\item{allel_c2}{Allele column, as string}
}
\value{
String vector with "C1/C1", "C1/C2", "C2/C2", or \code{NA_character_}
}
\description{
Takes two HLA-C alleles, classifies them individually into C1 or C2 using \code{\link{HLA_Classification}} and \code{\link{HLA_C_class_data}},
and returns the joint C1/C1, C1/C2 or C2/C2 class. If either allel classification is unknown or missing, returns NA.
}
\details{
There are only three outcome classes, C1/C1, C1/C2 and C2/C2. The location of each C-class before or after the "/" does not
match the input \code{allele_c1} and \code{_c2} - in other words if \code{allele_c1 = "C2"} and \code{allele_c2 = "C1"} the result is still C1/C2
The input alleles should be shortened to match the number of fields in the reference data.
}
\examples{
HLA_C_classification("01:02", "01:AWFCH")

dat <- data.frame(C1 = c("01:02", "02:03", "04:10"), C2 = c("01:AWFCH", "07:59", "05:50"))
HLA_C_classification(dat$C1, dat$C2)

# If either allele cannot be mapped to the reference data, NA is returned
dat_odd <- data.frame(C1 = c("01:02", "01:02", "01:02"), C2 = c(NA, "", "07:02:01"))
HLA_C_classification(dat_odd$C1, dat_odd$C2)
}
\seealso{
\code{\link{HLA_Classification}}, \code{\link{HLA_C_class_data}}
}
