% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_classification.R
\name{HLA_Classification}
\alias{HLA_Classification}
\title{Classify one HLA allele}
\usage{
HLA_Classification(allele, HLA_x_class)
}
\arguments{
\item{allele}{An allele string, including ":" characers but not the initial "C*"}

\item{HLA_x_class}{HLA classification reference data.frame (such as constructed by for example \code{HLA_C_class_load()}), where
the number of fields is the same as in \code{allele}}
}
\value{
String with  classification group from reference document, NA if unknown
}
\description{
Classifies an allele according to a reference dataframe. If an NMDP code is present, class is given if all possible
alleles are the same class, otherwise returns NA
}
\examples{
HLA_C_class <- HLA_C_class_load()
HLA_Classification("01:02", HLA_C_class)
HLA_Classification("01:AWFCH", HLA_C_class)
}
\seealso{
\code{\link{HLA_C_class_data}}, \code{\link{NMDP}}
}
