% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR3DL1_3DS1_assignment}
\alias{KIR3DL1_3DS1_assignment}
\title{Classifies KIR 3DL1 & 3DS1 alleles into KIR3DL1-H, -L, -N}
\usage{
KIR3DL1_3DS1_assignment(KIR3DL1_string, KIR3DS1_string)
}
\arguments{
\item{KIR3DL1_string}{KIR3DL1 string with first fields only, i.e. "001/095+004". Can contain "NEG" and "POS".}

\item{KIR3DS1_string}{KIR3DS1 string in same format}
}
\value{
String with "KIR3DL1-H", "KIR3DL1-L", "KIR3DL1-N" or "unknown"
}
\description{
Uses the categorization from \code{\link{ASSIGN_KIR3DL1}}
}
\examples{

KIR3DL1_3DS1_assignment("001/095+004", "NEG")
KIR3DL1_3DS1_assignment("POS", "NEG")
KIR3DL1_3DS1_assignment("005", "013/107")
}
\seealso{
\code{\link{ASSIGN_KIR3DL1}}
}
