% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kir.R
\name{KIR_present}
\alias{KIR_present}
\title{KIR present/absent}
\usage{
KIR_present(string)
}
\arguments{
\item{string}{KIR string}
}
\value{
TRUE if the KIR is present, FALSE if it's not present, NA if passed an empty/NA string
}
\description{
Determine whether the KIR is present, given a KIR string. Assumes an empty string, or a string with only "NEG" means the KIR is 
absent, any other results means it's present - make sure the data follows this assumption before passing to this function.
}
\examples{
KIR_present("003/034+003/034") # TRUE = present
KIR_present("NEG") # FALSE = absent
KIR_present("") # = NA
}
