% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMDP}
\alias{NMDP}
\title{NMDP allele codes}
\format{
\describe{
  \item{nmdp_codes}{NMDP code}
  \item{Allele}{Component alleles, separated by "/"}
}
}
\source{
\url{https://bioinformatics.bethematchclinical.org/hla-resources/allele-codes/allele-code-lists/allele-code-list-in-alphabetical-order/}
}
\usage{
NMDP
}
\description{
Lists possible alleles for NMDP codes
}
\keyword{datasets}
