% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munging.R
\name{shorten_allele}
\alias{shorten_allele}
\title{Reduce allele string to specified number of fields}
\usage{
shorten_allele(allele, fields = 2)
}
\arguments{
\item{allele}{String, including ":" but not initial "A*", so for example "01:02:01:03"}

\item{fields}{Number of fields to keep, default = 2 which would return "01:02"}
}
\value{
String with shorterned fields
}
\description{
Reduce a long allele string (i.e. "01:02:01:03g") down to fewer fields,
default being 2 fields. NMDP codes are preserved, trailing "g" are removed.
}
\examples{
shorten_allele("01:02:01:03")
shorten_allele("03:ADJRE")
shorten_allele("01:02:01:03", fields = 3)
}
